ExceptionHandling done

+++ employee_registration/src/main/java/com/employee/service/EmpSrvcImpl.java
@@ -6,12 +6,18 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.employee.bean.EmployeeUpdate;
 import com.employee.bean.Employeedto;
 import com.employee.entity.Employee;
+import com.employee.exception.NoNameResourceFoundException;
+import com.employee.exception.NoPincodeResourceFoundException;
+import com.employee.exception.NoSernameResourceFoundException;
+import com.employee.exception.ResourceAlreadyExistException;
+import com.employee.exception.ResourceNotFoundException;
 import com.employee.repository.IEmployeeRegistery;
 
 @Service
@@ -28,6 +34,9 @@ public class EmpSrvcImpl implements IEmployeeService {
 	//new Employee
 	@Transactional(readOnly = false)
 	public Employee registerEmp(Employee employee) {
+		if(iEmployeeRegistery.existsByid(employee.getId())) {
+			throw new ResourceAlreadyExistException("User "+employee.getId()+" already Existed");
+		}
 		return iEmployeeRegistery.save(employee);	
 	}
 	
@@ -35,17 +44,17 @@ public class EmpSrvcImpl implements IEmployeeService {
 	@Transactional(readOnly = true)
 	public List<Employeedto> getAllEmployee(){
 		
-		employeedtos = new ArrayList<Employeedto>();
 		employees = iEmployeeRegistery.findAll();
 		return listEmpTodto(employees);
 	}
 	
 	//get employee
 	@Transactional(readOnly = true)
-	public Employeedto getEmployee(int employeeId) {	
-	
-		//employee = iEmployeeRegistery.findById(employeeId);
+	public Employeedto getEmployee(int employeeId)  {	
+		if(iEmployeeRegistery.existsByid(employeeId))
 		employee = iEmployeeRegistery.findByid(employeeId);
+		else
+			throw new ResourceNotFoundException("given user "+employeeId+" not available");
 		return empTodto(employee);
 	}
 	
@@ -53,50 +62,77 @@ public class EmpSrvcImpl implements IEmployeeService {
 	@Transactional(readOnly = false)
 	public Employeedto updateEmployee(Employeedto employeedto) {
 		
-		 employee = iEmployeeRegistery.findByid(employeedto.getId());
-		 employee = dtoToemp(employeedto, employee);
-		 employee = iEmployeeRegistery.save(employee);
-		 return empTodto(employee);
+		if(iEmployeeRegistery.existsByid(employeedto.getId())) {
+			 employee = dtoToemp(employeedto, employee);
+			 employee = iEmployeeRegistery.save(employee);
+		}
+		else
+			throw new ResourceNotFoundException("given user "+employeedto.getId()+" not available");
+		return empTodto(employee);
 	}
 	
 	//soft delete employee
 	@Transactional(readOnly = false)
-	public void deleteEmployee(int employeeId) {
-		iEmployeeRegistery.deleteById(employeeId);
+	public void deleteEmployee(int employeeId) throws ResourceNotFoundException {
+		if(iEmployeeRegistery.existsByid(employeeId))
+     		iEmployeeRegistery.deleteById(employeeId);
+		else
+			throw new ResourceNotFoundException("given user "+employeeId+" not available");
 	}
 	
 	//hard delete employee
 	@Transactional(readOnly = false)
-	public void hardDeleteEmployee(int employeeId) {
-		iEmployeeRegistery.hardDeleteByid(employeeId);
+	public void hardDeleteEmployee(int employeeId) throws ResourceNotFoundException {
+		if(iEmployeeRegistery.existsByid(employeeId))
+	    	iEmployeeRegistery.hardDeleteByid(employeeId);
+		else
+			throw new ResourceNotFoundException("given user "+employeeId+" not available");
 	}
 	
 	//search by firstName
 	@Transactional(readOnly = true)
 	public List<Employeedto> searchByfirstName(String firstName){
-		List<Employee> employees;
-		employees = iEmployeeRegistery.findByfirstName(firstName);
+		
+		if(iEmployeeRegistery.existsByfirstName(firstName))
+	    	employees = iEmployeeRegistery.findByfirstName(firstName);
+		else
+			throw new NoNameResourceFoundException("given users by Name "+firstName+" is not available");
 		return listEmpTodto(employees);
 	}
 	
 	//search by lastName
 	@Transactional(readOnly = true)
 	public List<Employeedto> searchBylastName(String lastName){
-		List<Employee> employees;
-		employees = iEmployeeRegistery.findBylastName(lastName);
+		
+		if(iEmployeeRegistery.existsBylastName(lastName))
+	    	employees = iEmployeeRegistery.findBylastName(lastName);
+		else
+			throw new NoSernameResourceFoundException("given users by Sername "+lastName+" is not available");
 		return listEmpTodto(employees);
 	}
 	
 	//search by pincode
 	@Transactional(readOnly = true)
 	public List<Employeedto> searchByPincode(int pincode){
-		List<Employee> employees;
-		employees = iEmployeeRegistery.findBypincode(pincode);
+		
+		if(iEmployeeRegistery.existsBypincode(pincode))
+		    employees = iEmployeeRegistery.findBypincode(pincode);
+		else
+			throw new NoPincodeResourceFoundException("given users by Pincode "+pincode+" is not available");
+		return listEmpTodto(employees);
+	}
+	
+	//sort by dob/doj
+	@Transactional(readOnly = true)
+	public List<Employeedto> sortByField(String field){
+	
+		employees = iEmployeeRegistery.findAll(Sort.by(Sort.Direction.ASC, field));
 		return listEmpTodto(employees);
 	}
 	
 	//converting employee to dto
 	public Employeedto empTodto(Employee employee) {
+		
 		employeedto = new Employeedto();
 		employeedto.setFirstName(employee.getFirstName());
 		employeedto.setLastName(employee.getLastName());
@@ -113,7 +149,8 @@ public class EmpSrvcImpl implements IEmployeeService {
 	
 	//converting List employee to List dto
 	public List<Employeedto> listEmpTodto(List<Employee> employees){
-	return employees.stream().map(emp -> {
+	
+		return employees.stream().map(emp -> {
 			employeedto = new Employeedto();
 			employeedto.setFirstName(emp.getFirstName());
 			employeedto.setLastName(emp.getLastName());
